+++ jpf-domain/src/main/java/com/joiest/jpf/common/po/PayShopBatchCoupon.java
@@ -1,6 +1,7 @@
 package com.joiest.jpf.common.po;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class PayShopBatchCoupon implements Serializable {
@@ -51,7 +52,7 @@ public class PayShopBatchCoupon implements Serializable {
     /**
      * 对应豆数量
      */
-    private Integer dou;
+    private BigDecimal dou;
 
     /**
      * 是否激活 0=未激活 1=已激活
@@ -189,11 +190,11 @@ public class PayShopBatchCoupon implements Serializable {
         this.money = money;
     }
 
-    public Integer getDou() {
+    public BigDecimal getDou() {
         return dou;
     }
 
-    public void setDou(Integer dou) {
+    public void setDou(BigDecimal dou) {
         this.dou = dou;
     }
